requirements.txt
asgiref==3.6.0
Django==4.2
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pytz==2023.3
sqlparse==0.4.3
psycopg2==2.9.6
python-dotenv==1.0.0
black==23.3.0
drf-yasg==1.21.5
django-celery==3.3.1
django-celery-beat==2.5.0
redis==4.5.5
